fix: #33 adjusting file names and removing code smells
