+++ src/components/action-buttons/delete-button-homologation/index.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useState } from 'react';
+import { useState } from 'react';
 import { FaTrash } from 'react-icons/fa';
 import {
   Button,
