+++ src/config/routes/Routes.tsx
@@ -9,7 +9,7 @@ import { ListaProblemas } from '@/pages/categorias/problemas';
 import { RequireAuth } from '@/config/routes/require-auth';
 import { DefaultLayout } from '@/components/layout/default-layout';
 import { RegistrarChamado } from '@/pages/chamados/registrar';
-import { EditarChamadoExterno } from '@/pages/homologacao/editar-atendimentos-externos';
+import { EditarChamadoExterno } from '@/pages/homologacao/editar-agendamentos-externos';
 import { Agendamentos } from '@/pages/agendamentos';
 import { ScheduleExport } from '@/pages/exportacao_agendamentos';
 import { Tutoriais } from '@/pages/tutoriais';
