+++ src/features/homologations/components/issue-open-form/create-issue-form-open.tsx
@@ -14,7 +14,7 @@ import { MdEmail } from 'react-icons/md';
 import { ControlledSelect } from '@/components/form-fields';
 import { Input } from '@/components/form-fields/input';
 import { useGetAllCities } from '@/features/cities/api/get-all-cities';
-import { usePostCreateIssueOpen } from '@/features/issues/api/post-create-issue-open';
+import { usePostCreateIssueOpen } from '@/features/homologations/api/post-create-issue-open';
 import {
   IssueOpen,
   IssuePayloadOpen,
